+++ tests/e2e/setup.go
@@ -1,4 +1,4 @@
-package integration
+package e2e
 
 import (
 	"archive/tar"
